backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
pandas==2.0.1
pymc==5.3.1
pymc-marketing==0.10.0
arviz==0.15.1
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
xarray==0.19.0
netCDF4==1.5.7
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
flask
requests
python-dotenv
alembic